remove old page component
