+++ public/app/core/components/Page/Page.test.tsx
@@ -1,23 +1,28 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
 import { TestProvider } from 'test/helpers/TestProvider';
+import { getGrafanaContextMock } from 'test/mocks/getGrafanaContextMock';
 
-import { NavModelItem } from '@grafana/data';
+import { NavModelItem, PageLayoutType } from '@grafana/data';
 import { config } from '@grafana/runtime';
+import { HOME_NAV_ID } from 'app/core/reducers/navModel';
 
 import { Page } from './Page';
 import { PageProps } from './types';
 
 const pageNav: NavModelItem = {
-  text: 'Main title',
+  text: 'pageNav title',
   children: [
-    { text: 'Child1', url: '1', active: true },
-    { text: 'Child2', url: '2' },
+    { text: 'pageNav child1', url: '1', active: true },
+    { text: 'pageNav child2', url: '2' },
   ],
 };
-
 const setup = (props: Partial<PageProps>) => {
   config.bootData.navTree = [
+    {
+      id: HOME_NAV_ID,
+      text: 'Home',
+    },
     {
       text: 'Section name',
       id: 'section',
@@ -29,13 +34,17 @@ const setup = (props: Partial<PageProps>) => {
     },
   ];
 
-  return render(
-    <TestProvider>
+  const context = getGrafanaContextMock();
+
+  const renderResult = render(
+    <TestProvider grafanaContext={context}>
       <Page {...props}>
         <div data-testid="page-children">Children</div>
       </Page>
     </TestProvider>
   );
+
+  return { renderResult, context };
 };
 
 describe('Render', () => {
@@ -51,14 +60,25 @@ describe('Render', () => {
   it('should render header when pageNav supplied', async () => {
     setup({ pageNav });
 
-    expect(screen.getByRole('heading', { name: 'Main title' })).toBeInTheDocument();
+    expect(screen.getByRole('heading', { name: 'pageNav title' })).toBeInTheDocument();
     expect(screen.getAllByRole('tab').length).toBe(2);
   });
 
-  it('should get header nav model from redux navIndex', async () => {
-    setup({ navId: 'child1' });
+  it('should update chrome with section, pageNav and layout', async () => {
+    const { context } = setup({ navId: 'child1', pageNav, layout: PageLayoutType.Canvas });
+    expect(context.chrome.state.getValue().sectionNav.node.id).toBe('child1');
+    expect(context.chrome.state.getValue().pageNav).toBe(pageNav);
+    expect(context.chrome.state.getValue().layout).toBe(PageLayoutType.Canvas);
+  });
 
-    expect(screen.getByRole('heading', { name: 'Section name' })).toBeInTheDocument();
-    expect(screen.getAllByRole('tab').length).toBe(2);
+  it('should update document title', async () => {
+    setup({ navId: 'child1', pageNav });
+    expect(document.title).toBe('pageNav title - Child1 - Section name - Grafana');
+  });
+
+  it('should not include hideFromBreadcrumb nodes in title', async () => {
+    pageNav.children![0].hideFromBreadcrumbs = true;
+    setup({ navId: 'child1', pageNav });
+    expect(document.title).toBe('pageNav title - Child1 - Section name - Grafana');
   });
 });
