+++ public/app/core/components/Page/PageContents.tsx
@@ -1,8 +1,6 @@
 // Libraries
-import { cx } from '@emotion/css';
 import React from 'react';
 
-// Components
 import PageLoader from '../PageLoader/PageLoader';
 
 interface Props {
@@ -12,5 +10,7 @@ interface Props {
 }
 
 export const PageContents = ({ isLoading, children, className }: Props) => {
-  return <div className={cx('page-container', 'page-body', className)}>{isLoading ? <PageLoader /> : children}</div>;
+  let content = className ? <div className={className}>{children}</div> : children;
+
+  return <>{isLoading ? <PageLoader /> : content}</>;
 };
