+++ public/app/core/components/Page/PluginPage.tsx
@@ -3,7 +3,7 @@ import React, { useContext } from 'react';
 import { PluginPageProps } from '@grafana/runtime';
 import { PluginPageContext } from 'app/features/plugins/components/PluginPageContext';
 
-import { Page } from '../Page/Page';
+import { Page } from './Page';
 
 export function PluginPage({ actions, children, info, pageNav, layout, renderTitle, subTitle }: PluginPageProps) {
   const context = useContext(PluginPageContext);
