+++ public/app/features/dashboard/components/DashboardPermissions/AccessControlDashboardPermissions.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { Permissions } from 'app/core/components/AccessControl';
-import { Page } from 'app/core/components/PageNew/Page';
+import { Page } from 'app/core/components/Page/Page';
 import { contextSrv } from 'app/core/core';
 import { AccessControlAction } from 'app/types';
